Move `shouldUseGooglePayBrand` to utils and use same function in shortcode
