+++ client/payment-request/utils/utils.js
@@ -99,3 +99,19 @@ export const getErrorMessageFromNotice = ( notice ) => {
 	div.innerHTML = notice.trim();
 	return div.firstChild ? div.firstChild.textContent : '';
 };
+
+/**
+ * Whether or not to use Google Pay branded button in Chrome.
+ *
+ * @return {boolean} Use Google Pay button in Chrome.
+ */
+export const shouldUseGooglePayBrand = () => {
+	const ua = window.navigator.userAgent.toLowerCase();
+	const isChrome =
+		/chrome/.test( ua ) &&
+		! /edge|edg|opr|brave\//.test( ua ) &&
+		'Google Inc.' === window.navigator.vendor;
+	// newer versions of Brave do not have the userAgent string
+	const isBrave = isChrome && window.navigator.brave;
+	return isChrome && ! isBrave;
+};
